+++ docs/changelog.md
@@ -15,6 +15,7 @@ pages](https://github.com/dalibo/ldap2pg/pulls?utf8=%E2%9C%93&q=is%3Apr%20is%3Am
 - Accepts an SQL query to list ignored roles. `postgres:blacklist` is renamed
   `postgres:roles_blacklist_query`. ldap2pg ensure backward compatibility.
 - Apply roles blacklist to LDAP results.
+- Generate unique comment per role instead of shared comment per rule.
 - Move `on_unexpected_dn` to `ldap` query.